+++ pyetheroll/constants.py
@@ -2,6 +2,7 @@ from enum import Enum
 
 ROUND_DIGITS = 2
 DEFAULT_GAS_PRICE_GWEI = 4
+DEFAULT_GAS_PRICE_WEI = int(DEFAULT_GAS_PRICE_GWEI * 1e9)
 
 
 class ChainID(Enum):